Removed 'receiveCanceled' & minor refactoring.
None of the events that RTG subscribes to are cancellable so
'receiveCanceled' was pointless. Also, the minor refactoring was mostly
so that I could get my head into the new system as the new handler
names, whilst totally fine, were making it difficult for me to see what
was going on, as I was so used to the other naming convention. (Sorry
@srs-bsns !!!)