Removed logging from RTGAPI#getRTGBiome
The original intention here was that these methods would only be called
in low-usage settings and not from within iterations where they would be
potentially called thousands of times per second during chunk
generation.

For the original intention, logging would have been appropriate, but for
such a high usage setting of calling it in BiomeAnalyser, etc., logging
is not appropriate as it is quite an expensive and laggy process and
just spams the log needlessly with redundant info. Unsupported biomes
are already logged once during initialisation and this is the only
logging that needs to be done.

Signed-off-by: srs-bsns <forfrdm@gmail.com>