Added /rtg command tree
There is currently only one subcommand, 'whereami', which will show
discrepancies between what the biome is in the chunk biome array and
what the BiomeProvider thinks the biome should be for the position that
the command user is at in the world.

Since this is a debug command, it requires admin permission level (4).

Signed-off-by: srs-bsns <forfrdm@gmail.com>