Some config clean-up
Removed oreGenEventCancellation as it is vestigial.

Removed rtgWorldTypeByDefault and its use in
EventHandlerClient#onGuiCreateWorld as it is problematic when trying to
configure another world type when RTG is loaded.
For example, upon exiting the Customize World Screen after configuring
the FLAT or CUSTOM world type, or any other customizable world type, the
world type selection will revert back to RTG as the event handler is
executed upon the GUI being displayed (again). This can lead to crashes
if a user tries to create the world with RTG selected as the generator
options string (that is still set from the other world type) may contain
entries that are illegal for RTG.
Displaying the WorldTypeMessageGUI should be sufficient in letting
users know they need to select the RTG world type to get RTG terrain.

Fixes #1326

Signed-off-by: srs-bsns <forfrdm@gmail.com>