Adjust how tree density is handled
Being that the global and biome-specific config settings are both
multipliers, there is no need for one to override the other. Simply
multiply them together and check that the result does not exceed the
global maximum. This change means that the biome config setting is set
to 1.0 by default, and not a negative value to 'disable' it.

Also clarify the config comments.

Signed-off-by: srs-bsns <forfrdm@gmail.com>